+++ History.md
@@ -1,3 +1,10 @@
+### 3.0 / 2019-04-17
+
+*   Added a Faraday request middleware.
+*   Replaced Hurley example with examples for the use of the Faraday
+    middleware.
+*   Added global Marlowe configuration.
+
 ### 2.0 / 2016-11-16
 
 *   Breaking change: the correlation header defaults to `X-Request-Id` instead